0.6.2 Alpha 2

+++ adbshell_alpha.py
@@ -3,9 +3,17 @@
 #   adbshell_alpha.py
 #          Core
 #       By : 神郭
-#  Version : 0.6.1 Alpha 
+#  Version : 0.6.2 Alpha 2
 import sys , os , platform , getopt , shutil , datetime
 import zipfile as zip
+if os.path.exists('adbshellpy_home.py') and os.path.exists('adbshellpy_libhelper.py') and os.path.exists('adbshellpy_libapkfile.py') ==False:
+    print('''
+    Error:Core library files are missing!(Or one of them is missing.)
+    The files are :adbshellpy_home.py, adbshellpy_libhelper.py, adbshellpy_libapkfile.py and adbshell.py or adbshell_alpha.py.
+    Please reinstall or download adbshllpy,then restart the program.
+    Project address:https://github.com/AEnjoy/adbshellpy/
+    ''')
+    sys.exit(1)
 try:import configparser,urllib.request 
 except: pass
 def errexit(arg): #异常信息
@@ -42,8 +50,8 @@ if sys.hexversion < 0x03060000:
 #else
 
 #默认设置BEGIN 可在adbshell.ini adbshell.py修改默认选项
-version='0.6.1.2 dev'
-builddate='2020-7-24 23:54:15'
+version='0.6.2 Alpha 2'
+builddate='2020-7-30 23:30:19'
 run=0
 p=platform.system()
 checkflag=True
@@ -55,7 +63,7 @@ adbfile=str(os.environ.get('adbfile'))
 adbinit=0
 #changes
 try:
-    f_=open("Changlog", "r")
+    f_=open("Changlog", "r",encoding='UTF-8')
     changes=f_.read()
     f_.close()
 except:changes='E:更新日志文件"Changlog"不存在,无法查看更新记录!'
@@ -117,7 +125,7 @@ class update():#bra=branch
         print('完整包升级完成,请重启 adbshellpy实例')
         return True
     def isnewversionavailable(self,b=''):
-        url='https://github.com/AEnjoy/adbshellpy/raw/'+self.bra+'/version'
+        url='https://hub.fastgit.org/AEnjoy/adbshellpy/raw/'+self.bra+'/version'
         try:
             urllib.request.urlretrieve(url,'version.txt')
         except:
@@ -175,6 +183,7 @@ def who():
     adb.devices() #First running,activing service.
     hand=os.popen(adbfile+' devices')
     hand.readline() #第一行需要跳过
+    clear()
     if len(deviceslist)==0: #第一次执行/没有设备/添加设备列表
         for b in hand:
             try:
@@ -288,18 +297,13 @@ class adbcommand():
     def root(self):
         self._adbc('root')
     def reboot(self,mode=0):#0 不带参数 1.-p 2.fastboot(bl) 3.recovery 4.sideload 5.挖煤
-        if mode == 0:
-            self._adbc('reboot')
-        if mode == 1:
-            self._adbc('reboot -p')
-        if mode == 2:
-            self._adbc('reboot bootloader')
-        if mode == 3:
-            self._adbc('reboot recovery')
-        if mode == 4:
-            self._adbc('reboot sideload')
-        if mode == 5:
-           self. _adbc('reboot download')
+        if mode == 0:self._adbc('reboot')
+        if mode == 1:self._adbc('reboot -p')
+        if mode == 2:self._adbc('reboot bootloader')
+        if mode == 3:self._adbc('reboot recovery')
+        if mode == 4:self._adbc('reboot sideload')
+        if mode == 5:self. _adbc('reboot download')
+        if mode == 6:self._adbc('reboot edl')
 
     def install(self,apkfile,command='-g -d'):
         self._adbc("install "+command+" "+'"'+apkfile+'"')
@@ -388,24 +392,16 @@ def setmode():#setmode parseinput(2)
     *您也可以通过手动编辑adbshell.ini来修改设置                                      *
     **********************************Setmode*****************************************
     ''')
-    try:import adbshellpy_home
-    except:
-        update().download_lib('adbshellpy_home')
-        import adbshellpy_home  
+    import adbshellpy_home  
     adbshellpy_home.parseinput(2)
 
 def Console():
-    clear()
     global run,adbinit
     if run == 0:
         if adbinit==0:adb.kill_server()#跳过运行时kill adb服务
         who()
         run=1
-    global nowdevice
-    try:import adbshellpy_home
-    except:
-        update().download_lib('adbshellpy_home')
-        import adbshellpy_home      
+    import adbshellpy_home      
     adbshellpy_home.home()
     adbshellpy_home.parseinput(1)
 
@@ -418,7 +414,7 @@ class _Options(object):
 
 def usage():
     print("""
-    用法:adbshell.py [apkfile(s)] [args or Console -MORE] 
+    用法:adbshell.py [apkfile(s)] [args or Console] [-MORE] 
     [apkfile(s)] [开发中]
     安装apk文件至手机 支持多个文件
     [args]
@@ -434,7 +430,7 @@ def usage():
     update       升级ADBSystemTOOLBOX程序,将访问GitHub获取升级
     environment  显示程序运行环境变量的设置及其它信息
     exit         退出程序
-    -MORE        程序内部的各种功能
+    [-MORE]        程序内部的各种功能
     """)
     errexit(2)
 
@@ -475,14 +471,9 @@ def ParseArguments(args): #解析参数
 
 def apkinstallmode(install=False):#开发计划2
     if install==True:
-        try:
-            import adbshellpy_libapkfile
-        except:
-            update().download_lib('adbshellpy_libapkfile')
-            import adbshellpy_libapkfile
+        import adbshellpy_libapkfile
         adbshellpy_libapkfile.mainex(_Options.apkfile)
         
-
 def main(args):
   global checkflag,branch,uselinuxpkgmanagertoinstalladb,adbfile,conf
   try:
@@ -493,10 +484,41 @@ def main(args):
   c=['shutdown','rec','bl','edl','sideload','download','install','uninstall','compile',
       'shell','root','start_server','kill_server','devices','tcpipconnect','usb','reboot',
       'disable','enable','clear','applist','pull','push','windowmode','input','settings',
-      'dumpsys','screencap','relatedapk','who'
+      'dumpsys','screencap','relatedapk','who','kfmark','icebox','update','changes','piebridge',
+      'shizuku'
      ]#内置命令
   if cmd in c:#开发计划3
-      pass
+      adbshellpy_home
+      fun=adbshellpy_home.func_()
+      if cmd=='shutdown':fun.shutdown()
+      if cmd=='kfmark':fun.kfmark()
+      if cmd=='rec':fun.rec()
+      if cmd=='bl':fun.bl()
+      if cmd=='edl':fun.edl()
+      if cmd=='sideload':fun.sideload()
+      if cmd=='download':fun.download()
+      if cmd=='install':fun.install()
+      if cmd=='uninstall':fun.uninstall()
+      if cmd=='compile':fun.compile()
+      if cmd=='shell':fun.shell()
+      if cmd=='reboot':fun.reboot()
+      if cmd=='disable':fun.disable()
+      if cmd=='enable':fun.enable()
+      if cmd=='clear':fun.clear()
+      if cmd=='applist':fun.applist()
+      if cmd=='pull':fun.pull()
+      if cmd=='push':fun.push()
+      if cmd=='windowmode':fun.windowmode()
+      if cmd=='input':fun.input()
+      if cmd=='dumpsys':fun.dumpsys()
+      if cmd=='screencap':fun.screencap()
+      if cmd=='relatedapk':fun.relatedapk()
+      if cmd=='icebox':fun.icebox()
+      if cmd=='update':fun.update()
+      if cmd=='changes':fun.changes_()
+      if cmd=='piebridge':fun.piebridge()
+      if cmd=='shizuku':fun.shizuku()
+      sys.exit(0)
   checkflag=opt.installcheck
   if sys.hexversion < 0x03060000:
       errexit(3)
